+++ src/main/java/roomescape/web/argumentresolver/LoginMemberArgumentResolver.java
@@ -9,7 +9,7 @@ import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.method.support.ModelAndViewContainer;
 import roomescape.dto.login.LoginMember;
 import roomescape.dto.token.TokenDto;
-import roomescape.infrastructure.AuthorizationExtractor;
+import roomescape.infrastructure.auth.AuthorizationExtractor;
 import roomescape.service.AuthService;
 
 @Component
